+++ src/async-slice.d.ts
@@ -7,23 +7,37 @@
  */
 
 import { AsyncInputIterable, AsyncGeneratorIterator } from './internal/async-iterable';
+declare function asyncSlice<T = any>(opts: {
+  readonly start?: number;
+  readonly end?: number;
+  readonly step?: number;
+}): (iterable: AsyncInputIterable<T>) => AsyncGeneratorIterator<T>;
 declare function asyncSlice<T = any>(
-  opts:
-    | number
-    | {
-        readonly start?: number;
-        readonly end?: number;
-        readonly step?: number;
-      },
+  start?: number,
+  end?: number,
+  step?: number,
 ): (iterable: AsyncInputIterable<T>) => AsyncGeneratorIterator<T>;
 declare function asyncSlice<T = any>(
-  opts:
-    | number
-    | {
-        readonly start?: number;
-        readonly end?: number;
-        readonly step?: number;
-      },
+  opts: {
+    readonly start?: number;
+    readonly end?: number;
+    readonly step?: number;
+  },
+  iterable: AsyncInputIterable<T>,
+): AsyncGeneratorIterator<T>;
+declare function asyncSlice<T = any>(
+  start: number,
+  iterable: AsyncInputIterable<T>,
+): AsyncGeneratorIterator<T>;
+declare function asyncSlice<T = any>(
+  start: number,
+  end: number,
+  iterable: AsyncInputIterable<T>,
+): AsyncGeneratorIterator<T>;
+declare function asyncSlice<T = any>(
+  start: number,
+  end: number,
+  step: number,
   iterable: AsyncInputIterable<T>,
 ): AsyncGeneratorIterator<T>;
 export default asyncSlice;